+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  * remove `FlMlKitScanningMethodCall()`
  * add `onCreateView` for `FlMlKitScanning()`
  * add `notPreviewed` for `FlMlKitScanning()`
+ * add `frequency` for `FlMlKitScanning()`
  * add `stopPreview()` for `FlMlKitScanningController()`
  * modify `onFlashChange` to `onFlashChanged`
  * modify `onZoomChange` to `onZoomChanged`